First version of getStatic and putStatic.
The current solution is veeery rough. It calls a method in C++ to get
the address of the field. As I don't want to stop the jit compiler, the
only generic enough information I can provide to this function is the
callee ClassFile and the callee reference index within such a ClassFile.
That it a least ugly (I better description is shit).
After that, I just move from/to the address